up code

+++ src/main/java/com/nahuannghia/shopnhn/repository/ProductCategoryRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Long> {
+public interface ProductCategoryRepository extends JpaRepository<ProductCategory, Integer> {
+    // Có thể thêm các phương thức tùy chỉnh nếu cần
 }
